==> notebooks/2.0-jpb-preprocessing.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split

pd.pandas.set_option('display.max_columns', None)

import warnings
warnings.simplefilter(action='ignore')

lending_data = pd.read_csv('lending_data.csv')

# clean features
def remove_percentage_sign(df, feat):
    df = df.copy()
    df[feat] = df[feat].str.replace(r'%', '').astype('float')
    return df

for feat in ['int_rate', 'revol_util']:
    lending_data = remove_percentage_sign(lending_data, feat)

lending_train, lending_val, rate_train, rate_val = train_test_split(
    lending_data, lending_data['int_rate'], test_size=0.1, random_state=12)

## Missing Values

### High Percentage

def feats_with_na_above_thresh(df, min_thresh, max_thresh=1.0):
    missing_feats = [feat for feat in df.columns if 
                     (df[feat].isnull().sum() / df.shape[0]) > min_thresh and
                     df[feat].isnull().sum() / df.shape[0] <= max_thresh]
    return df[missing_feats].isnull().mean().sort_values(ascending=False)

high_perc_feats = feats_with_na_above_thresh(lending_train, 0.9)

# drop features that have > 90% missing in training data
lending_train = lending_train.drop(high_perc_feats.index, axis=1)
lending_val = lending_val.drop(high_perc_feats.index, axis=1)

### Categorical Variables

cat_feats_with_na = [
    feat for feat in lending_train.columns if lending_train[feat].isnull().sum() > 0 and
        lending_train[feat].dtypes == 'O'
]

lending_train[cat_feats_with_na].isnull().mean()

# make a new label 'Missing' for the 2 features that have 7% missing
lending_train[['emp_title', 'emp_length']] = lending_train[
    ['emp_title', 'emp_length']].fillna('Missing')
lending_val[['emp_title', 'emp_length']] = lending_val[
    ['emp_title', 'emp_length']].fillna('Missing')

cat_feats_with_small_na = [
    feat for feat in cat_feats_with_na if lending_train[feat].isnull().sum() > 0
]

# drop data points w/ very little missing
lending_train = lending_train.dropna(subset=cat_feats_with_small_na)
lending_val = lending_val.dropna(subset=cat_feats_with_small_na)

# sanity checks
print([feat for feat in cat_feats_with_na if lending_train[feat].isnull().sum() > 0])
print([feat for feat in cat_feats_with_na if lending_val[feat].isnull().sum() > 0])

### Numerical Variables

num_feats_with_na = [
    feat for feat in lending_train.columns if lending_train[feat].isnull().sum() > 0 and
        lending_train[feat].dtypes != 'O'
]

lending_train[num_feats_with_na].isnull().mean()

With features that have greater than 5% of instances missing, we will create a binary missing value indicator variable, and replace the missing values in original variable with the median.

for feat in [
    'mths_since_last_delinq', 'mths_since_last_record', 'mths_since_last_major_derog',
    'il_util', 'mths_since_recent_bc_dlq', 'mths_since_recent_inq', 
    'mths_since_recent_revol_delinq', 'num_tl_120dpd_2m']:
    median_val = lending_train[feat].median()
    lending_train[feat+'_na'] = np.where(lending_train[feat].isnull(), 1, 0)
    lending_val[feat+'_na'] = np.where(lending_val[feat].isnull(), 1, 0)
    lending_train[feat] = lending_train[feat].fillna(median_val)
    lending_val[feat] = lending_val[feat].fillna(median_val)

# drop the remaining instances
num_feats_with_small_na = [
    feat for feat in num_feats_with_na if lending_train[feat].isnull().sum() > 0
]

lending_train = lending_train.dropna(subset=num_feats_with_small_na)
lending_val = lending_val.dropna(subset=num_feats_with_small_na)

# sanity checks
print([feat for feat in num_feats_with_na if lending_train[feat].isnull().sum() > 0])
print([feat for feat in num_feats_with_na if lending_val[feat].isnull().sum() > 0])

## Categorical Variables

### Nonpredictive Variables

We will remove features that did not show any relationship with interest rates from the last notebook.

non_pred_cat_feats = [
    'emp_title', 'emp_length', 'issue_d', 'purpose', 'title', 'addr_state',
    'initial_list_status', 'zip_code', 'earliest_cr_line'
]

lending_train.drop(non_pred_cat_feats, axis=1, inplace=True)
lending_val.drop(non_pred_cat_feats, axis=1, inplace=True)

cat_feats = [
    feat for feat in lending_train.columns 
    if lending_train[feat].dtype == 'O'
]

### Rare Labels

We will replace all values of categorical variables that account for less than 1% of loans with 'Rare'.

def get_freq_labels(df, feat, tgt, rare_thresh):
    df = df.copy()
    cat_perc = df.groupby(feat)[tgt].count() / df.shape[0]
    return cat_perc[cat_perc > rare_thresh].index

for feat in cat_feats:
    freq_labels = get_freq_labels(lending_train, feat, 'int_rate', 0.01)
    lending_train[feat] = np.where(lending_train[feat].isin(
        freq_labels), lending_train[feat], 'Rare')
    lending_val[feat] = np.where(lending_val[feat].isin(
        freq_labels), lending_val[feat], 'Rare')

# sanity checks
print([feat for feat in lending_train.columns if lending_train[feat].isnull().sum() > 0])
print([feat for feat in lending_val.columns if lending_val[feat].isnull().sum() > 0])

### OHE

lending_train = pd.get_dummies(lending_train)
lending_val = pd.get_dummies(lending_val)

# sanity check
print(lending_train.shape)
print(lending_val.shape)

# save cleaned & processed dataframes
lending_train.to_csv('lending_train.csv', index=False)
lending_val.to_csv('lending_val.csv', index=False)
