==> notebooks/3.0-jpb-feature-selection.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from sklearn.feature_selection import SelectFromModel
from sklearn.linear_model import Lasso
from sklearn.ensemble import RandomForestRegressor

pd.pandas.set_option('display.max_columns', None)

import warnings
warnings.simplefilter(action='ignore')

lending_train = pd.read_csv('lending_train.csv')
lending_val = pd.read_csv('lending_val.csv')

print(lending_train.shape)
print(lending_val.shape)
lending_train.head()

rate_train = lending_train['int_rate']
rate_val = lending_val['int_rate']

lending_train.drop('int_rate', axis=1, inplace=True)
lending_val.drop('int_rate', axis=1, inplace=True)

## Feature Selection
### Lasso

We will do a rough feature selection using a Lasso linear model.

sel_ = SelectFromModel(Lasso(alpha=1.5, random_state=3))
sel_.fit(lending_train, rate_train)

sel_feats = lending_train.columns[(sel_.get_support())]
print(f'Number of selected features: {len(sel_feats)}')
print(sel_feats)

# save selected features
pd.Series(sel_feats).to_csv('lasso_sel_feats.csv', index=False)
